@@ -1,6 +1,6 @@
  
 import { Directive, Input, OnInit, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
-import { MatGridList } from '@angular/material';
+import { MatGridList } from '@angular/material/grid-list';
 import { MediaObserver, MediaChange } from '@angular/flex-layout';
 
 export interface IResponsiveColumnsMap {
